fix: do NOT tweak `v:errmsg`
What we did was probably useless.

Why did we do that originally?

We search for sth with no match.
Then we search for sth with a match.
The test will succeed, but there was no error this time.

So, why do we remove the code?
Because it's not an issue.
`search#nohls()` will be called while it's not needed.
There's no side effect.